requirements.txt
halo
openai=0.28
pyyaml